rag_practice_project/requirements.txt
# AI Client Framework - Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0

# OpenAI support
openai>=1.0.0
tiktoken>=0.5.0

# Gemini support
google-generativeai>=0.3.0

# Anthropic support (optional)
anthropic>=0.7.0

# Optional: For testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

pydantic>=2.5.2
pydantic-settings>=2.0.0  # Required by ChromaDB

langgraph
langchain-core
langsmith

numpy==1.26.4
sqlalchemy>=2.0.0
python-dateutil>=2.8.0

pandas==2.2.2
kagglehub
sentence-transformers
torch
chromadb>=0.5.0
scikit-learn
rouge-score
bert-score
matplotlib
seaborn
plotly
tqdm
pyyaml
pyarrow # For Parquet support